refactor: convert 7 [param] routes to query params, fix SuperTokens init timing
- Convert /practitioner/[id], /clinic/[id], /profile/[path],
  /assessments/[id], /exercise/[id], /schedule/[id], /record/[id]
  to query-param based routing with dispatcher pattern
- Remove 3 Next.js API routes (provinces/cities/districts),
  call wilayah.id directly from client
- Fix SuperTokens init race: defer to useEffect after RuntimeConfig
  resolves, pass correct appInfo from context to frontendConfig
- Fix getAppInfo() fallback: use window.location.origin instead of empty string
- Add /api/v1/auth/ to CSRF exempt prefixes for auth calls
- Update all navigation links (Link href, router.push) to use query-param format
- Remove dead book-practitioner sub-route
