--- src/app/api/cities/[provinceCode]/route.ts
@@ -1,20 +1,0 @@
-import axios from 'axios';
-import { NextRequest, NextResponse } from 'next/server';
-
-export async function GET(req: NextRequest) {
-  const url = new URL(req.url);
-  const provinceCode = url.pathname.split('/').pop();
-  try {
-    const response = await axios.get(
-      `https://wilayah.id/api/regencies/${provinceCode}.json`
-    );
-    return NextResponse.json(response.data);
-  } catch (error: any) {
-    console.error('Error fetching cities:', error.message);
-
-    return NextResponse.json(
-      { message: 'Error fetching cities' },
-      { status: 500 }
-    );
-  }
-}
