+++ src/app/exercise/page.tsx
@@ -1,20 +1,16 @@
-import BackButton from '@/components/general/back-button';
-import Header from '@/components/header';
-import NavigationBar from '@/components/navigation-bar';
-import ExcerciseList from './excercise-list';
+'use client';
 
-export default function Exercise() {
-  return (
-    <>
-      <NavigationBar />
-      <Header showChat={false}>
-        <div className='flex w-full items-center'>
-          <BackButton route='/' />
+import { useSearchParams } from 'next/navigation';
+import ExerciseDetail from './exercise-detail';
+import ExerciseList from './exercise-list';
 
-          <div className='text-[14px] font-bold text-white'>Self Excercise</div>
-        </div>
-      </Header>
-      <ExcerciseList />
-    </>
-  );
+export default function ExercisePage() {
+  const searchParams = useSearchParams();
+  const exerciseId = searchParams.get('exerciseId');
+
+  if (exerciseId) {
+    return <ExerciseDetail />;
+  }
+
+  return <ExerciseList />;
 }
