+++ src/app/home-content-clinician.tsx
@@ -208,7 +208,7 @@ export default function HomeContentClinician() {
       </div>
 
       <div className='mt-2 flex flex-col gap-2 py-4'>
-        <Link href={`assessments/${selectedAssessment?.id}`}>
+        <Link href={`/assessments?assessmentsId=${selectedAssessment?.id}`}>
           <Button className='bg-secondary h-full w-full rounded-xl p-4 text-white'>
             Start Test
           </Button>
