+++ src/app/record/patient-record.tsx
@@ -1,4 +1,5 @@
 'use client';
+/* eslint-disable max-lines */
 
 import Avatar from '@/components/general/avatar';
 import BackButton from '@/components/general/back-button';
@@ -194,7 +195,7 @@ export default function PatientRecord() {
                 handleSetRecordFilter('query', event.target.value)
               }
               placeholder='Search Entry & Record'
-              className='mr-4 h-[50px] w-full border-0 bg-[#F9F9F9] text-primary'
+              className='text-primary mr-4 h-[50px] w-full border-0 bg-[#F9F9F9]'
               startIcon={<SearchIcon className='text-[#ABDCDB]' width={16} />}
             />
             <RecordFilter
@@ -209,7 +210,7 @@ export default function PatientRecord() {
 
           <div className='flex gap-4'>
             {recordFilter.start_date && recordFilter.end_date && (
-              <Badge className='mt-4 rounded-md bg-secondary px-4 py-[3px] font-normal text-white'>
+              <Badge className='bg-secondary mt-4 rounded-md px-4 py-[3px] font-normal text-white'>
                 {recordFilter.start_date == recordFilter.end_date
                   ? format(recordFilter.start_date, 'dd MMM yy')
                   : format(recordFilter.start_date, 'dd MMM yy') +
@@ -218,7 +219,7 @@ export default function PatientRecord() {
               </Badge>
             )}
             {filterTypeLabel && (
-              <Badge className='mt-4 rounded-md bg-secondary px-4 py-[3px] font-normal text-white'>
+              <Badge className='bg-secondary mt-4 rounded-md px-4 py-[3px] font-normal text-white'>
                 {filterTypeLabel}
               </Badge>
             )}
@@ -237,10 +238,10 @@ export default function PatientRecord() {
               alt='writing'
             />
             <div className='ml-2 flex flex-col'>
-              <span className='text-[12px] font-bold text-primary'>
+              <span className='text-primary text-[12px] font-bold'>
                 Start Writting
               </span>
-              <span className='text-[10px] text-primary'>
+              <span className='text-primary text-[10px]'>
                 Express your current feelings
               </span>
             </div>
@@ -286,7 +287,7 @@ export default function PatientRecord() {
                 category: typeMappings[record.type]?.category,
                 title
               }).toString();
-              const url = `record/${recordId}?${queryParams}`;
+              const url = `/record?recordId=${recordId}&${queryParams}`;
 
               const { displayName, email } = getPractitionerInfo(record);
               const { initials, backgroundColor } = generateAvatarPlaceholder({
@@ -316,7 +317,7 @@ export default function PatientRecord() {
                       <div className='text-[12px] font-bold'>
                         {formattedTitle}
                       </div>
-                      <div className='line-clamp-3 overflow-hidden text-ellipsis text-[10px]'>
+                      <div className='line-clamp-3 overflow-hidden text-[10px] text-ellipsis'>
                         <ReactMarkdown components={customMarkdownComponents}>
                           {cleanDescription}
                         </ReactMarkdown>
