+++ src/app/record/practitioner-record-soap.tsx
