+++ src/app/record/practitioner-record.tsx
@@ -1,4 +1,5 @@
 'use client';
+/* eslint-disable max-lines */
 
 import Avatar from '@/components/general/avatar';
 import BackButton from '@/components/general/back-button';
@@ -217,7 +218,7 @@ export default function PractitionerRecord() {
                 handleSetRecordFilter('query', event.target.value)
               }
               placeholder='Search Entry & Record'
-              className='mr-4 h-[50px] w-full border-0 bg-[#F9F9F9] text-primary'
+              className='text-primary mr-4 h-[50px] w-full border-0 bg-[#F9F9F9]'
               startIcon={<SearchIcon className='text-[#ABDCDB]' width={16} />}
             />
             <RecordFilter
@@ -232,7 +233,7 @@ export default function PractitionerRecord() {
 
           <div className='flex gap-4'>
             {recordFilter.start_date && recordFilter.end_date && (
-              <Badge className='mt-4 rounded-md bg-secondary px-4 py-[3px] font-normal text-white'>
+              <Badge className='bg-secondary mt-4 rounded-md px-4 py-[3px] font-normal text-white'>
                 {recordFilter.start_date == recordFilter.end_date
                   ? format(recordFilter.start_date, 'dd MMM yy')
                   : format(recordFilter.start_date, 'dd MMM yy') +
@@ -241,7 +242,7 @@ export default function PractitionerRecord() {
               </Badge>
             )}
             {filterTypeLabel && (
-              <Badge className='mt-4 rounded-md bg-secondary px-4 py-[3px] font-normal text-white'>
+              <Badge className='bg-secondary mt-4 rounded-md px-4 py-[3px] font-normal text-white'>
                 {filterTypeLabel}
               </Badge>
             )}
@@ -260,10 +261,10 @@ export default function PractitionerRecord() {
               alt='writing'
             />
             <div className='ml-2 flex flex-col'>
-              <span className='text-[12px] font-bold text-primary'>
+              <span className='text-primary text-[12px] font-bold'>
                 SOAP Report
               </span>
-              <span className='text-[10px] text-primary'>Start Writting</span>
+              <span className='text-primary text-[10px]'>Start Writting</span>
             </div>
           </Link>
         </div>
@@ -327,7 +328,7 @@ export default function PractitionerRecord() {
                 category: typeMappings[record.type]?.category,
                 title
               }).toString();
-              const url = `record/${recordId}?${queryParams}`;
+              const url = `/record?recordId=${recordId}&${queryParams}`;
 
               const { displayName, email } = getPractitionerInfo(record);
               const { initials, backgroundColor } = generateAvatarPlaceholder({
@@ -357,7 +358,7 @@ export default function PractitionerRecord() {
                       <div className='text-[12px] font-bold'>
                         {formattedTitle}
                       </div>
-                      <div className='line-clamp-3 overflow-hidden text-ellipsis text-[10px]'>
+                      <div className='line-clamp-3 overflow-hidden text-[10px] text-ellipsis'>
                         <ReactMarkdown components={customMarkdownComponents}>
                           {cleanDescription}
                         </ReactMarkdown>
