+++ src/components/supertokensProvider.tsx
@@ -1,16 +1,25 @@
 'use client';
+import { RuntimeConfigContext } from '@/components/general/runtime-config-provider';
 import { frontendConfig, setRouter } from '@/config/frontendConfig';
 import { usePathname, useRouter } from 'next/navigation';
-import React from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import SuperTokensReact, { SuperTokensWrapper } from 'supertokens-auth-react';
 
-if (typeof window !== 'undefined') {
-  SuperTokensReact.init(frontendConfig());
-}
-
-export const SuperTokensProviders: React.FC<React.PropsWithChildren<{}>> = ({
-  children
-}) => {
+export const SuperTokensProviders: React.FC<
+  React.PropsWithChildren<object>
+> = ({ children }) => {
+  const runtimeConfig = useContext(RuntimeConfigContext);
+  const [initDone, setInitDone] = useState(false);
   setRouter(useRouter(), usePathname() || window.location.pathname);
+
+  useEffect(() => {
+    if (runtimeConfig && !initDone) {
+      SuperTokensReact.init(frontendConfig(runtimeConfig.appInfo));
+      setInitDone(true);
+    }
+  }, [runtimeConfig, initDone]);
+
+  if (!initDone) return null;
+
   return <SuperTokensWrapper>{children}</SuperTokensWrapper>;
 };
