+++ src/services/api/cities.tsx
@@ -1,11 +1,13 @@
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 
+const BASE = 'https://wilayah.id/api';
+
 export const useGetProvinces = () => {
   return useQuery({
     queryKey: ['provinces'],
     queryFn: async () => {
-      const response = await axios.get('/api/provinces');
+      const response = await axios.get(`${BASE}/provinces.json`);
       return response.data.data;
     },
     select: response => response || []
@@ -17,8 +19,10 @@ export const useGetCities = (provinceCode: number) => {
     queryKey: ['cities', provinceCode],
     queryFn: async () => {
       if (provinceCode === 0) return null;
-      const response = await axios.get(`/api/cities/${provinceCode}`);
-      return response.data.data;
+      const response = await axios.get(
+        `${BASE}/regencies/${provinceCode}.json`
+      );
+      return response.data;
     },
     enabled: provinceCode !== undefined && provinceCode !== null,
     select: response => response || null
@@ -30,8 +34,8 @@ export const useGetDistricts = (cityCode: number) => {
     queryKey: ['districts', cityCode],
     queryFn: async () => {
       if (cityCode === 0) return null;
-      const response = await axios.get(`/api/districts/${cityCode}`);
-      return response.data.data;
+      const response = await axios.get(`${BASE}/districts/${cityCode}.json`);
+      return response.data;
     },
     enabled: cityCode !== undefined && cityCode !== null,
     select: response => response || null
